check for returned value from ping and write default errro

+++ check_neighs.sh
@@ -6,7 +6,7 @@ PINGFOLDER=/var/www/html/ping
 PINGFILE=$PINGFOLDER/ping.csv
 PINGFILE_BAK=/tmp/ping.csv.bak
 TXTINFO_URL="http://localhost:2006/links"
-MAX_LINES=1000
+MAX_LINES=100
 
 mkdir -p $PINGFOLDER
 
@@ -14,8 +14,10 @@ mkdir -p $PINGFOLDER
 wget -q -O $NEIGHS $TXTINFO_URL 
 NOW=`date +"%Y-%m-%d %H:%M"`
 
-tail -n $MAX_LINES $PINGFILE > $PINGFILE_BAK
-mv $PINGFILE_BAK $PINGFILE
+if  [ -e $PINGFILE ]; then
+ tail -n $MAX_LINES $PINGFILE > $PINGFILE_BAK
+ mv $PINGFILE_BAK $PINGFILE
+fi 
 
 LINKS=`cat $NEIGHS | grep -v IP | grep -v "Table"`;
 IFS=$'\n'
@@ -23,11 +25,16 @@ for link in $LINKS; do
  echo $link
  SIP=`echo $link |  cut -f 1`
  DIP=`echo $link |  cut -f 2`
- $PINGRES=`ping -c $PINGS -q $DIP | grep "max" | cut -d= -f 2 | sed "s/\// /g" | sed "s/ms//g"` 
- if [ $? -ne 0 ]; then
-    echo -n "$NOW $SIP $DIP" >> $PINGFILE;
+ # ping, then get the stats and strip newline
+ PINGRES=$(ping -c $PINGS -q $DIP | grep "max" | cut -d= -f 2 | \
+		sed "s/\// /g" | sed "s/ms//g" | sed "s/\n//g")
+ echo -n "$NOW $SIP $DIP" >> $PINGFILE;
+ if [ -n "$PINGRES" ]; then # check if string is not null
     echo $PINGRES >> $PINGFILE;
+ else
+    echo " -1 -1 -1" >> $PINGFILE;
  fi
+
 done;
 IFS=$' '
